Additionally support Bun
Bun runs a lot faster (current test suite goes from ~4s by node to ~60ms
by bun) so add a bun lockfile. This works without additional config
files, which surprised me!

Due to incomplete Jest support, this has required a hacky workaround when
trying to establish whether the current file is being run as part of a
test or not (i.e. so that we do not attempt to run against contest
inputs in the test suite).